requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
discord[voice]
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
discord==2.3.2
discord.py==2.4.0
dnspython==2.7.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.1
pymongo==4.11.3
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
tweepy==4.15.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
yarl==1.18.3
jinja2 
pillow